Refactor the whole enchilada
git-smartlist is only useful when paired with a "git log" equivalent,
and while this pairing can be done with git complex aliases, there's no
need for everyone to setup the same alias,

Make git-smartlist receive the git command to be executed.

My only problem with this approach before was that gitk isn't a command,
but that can be easily fixed with complex aliases like:

  v = "!f() { gitk $@ & }; f"

This simplifies everything, and there's no need for external functions
nor aliases.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>